backend/python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
PyYAML==6.0.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
google-generativeai==0.8.3
email-validator==2.1.1
